add : SplitterTest
